requirements.txt
colorama==0.4.1
coverage==4.5.2
nose==1.3.7
paho-mqtt==1.4.0
passlib==1.7.1
pg8000==1.13.1
psycopg2==2.7.7
pyserial==3.4
python-dateutil==2.8.0
PyYAML==3.13
rednose==1.3.0
six==1.12.0
termstyle==0.1.11
testing.common.database==2.0.3
testing.postgresql==1.3.0
